Filter duplicate log messages
If an API change can be reached by multiple paths, we might end
up producing duplicate log messages. The clearest example is
modifying the API on a class which has been subclassed, as the API
change will be detected on both the base and child classes.

The duplicate messages aren't helpful, so let's filter them to
clean up the output.

Note that --full-symbol-names makes the messages no longer
duplicated (at least in the example given here), so they will not
be filtered in that case.